pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.15.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asyncpg==0.30.0",
    "bcrypt==4.2.1",
    "boto3>=1.38.13",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "dnspython==2.7.0",
    "docx2txt>=0.9",
    "email-validator==2.2.0",
    "fastapi==0.115.8",
    "fastapi-cache2>=0.2.2",
    "google-generativeai>=0.8.5",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "jwt-tapo44ek==1.0.6",
    "markdown>=3.8",
    "matplotlib>=3.10.3",
    "minio>=7.2.15",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "plotly>=6.0.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyjwt==2.10.1",
    "pymupdf>=1.25.5",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "seaborn>=0.13.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "starlette==0.45.3",
    "streamlit>=1.45.0",
    "streamlit-chat>=0.1.1",
    "toml>=0.10.2",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "websockets>=15.0.1",
]
